Update exp.lua
Allowed -bexp in c, so fixed the parameter handling over to lua for this negative bexp. This could be used for worldboss events or cursed hours, temporary reducing exp. of course this exp punish should be comnpared by a kind of special loot bonus or a big exp bonus time after this punish.

Changed the output for exp/bexp to a more understandable textflow.